Split AST statement functions
